Fixed frontend format/add LoginPage to backend
I changed the frontend format so that the sandbox wasn't the main function rather made it a subfile with the main and index being the main one. I add a bunch of temporary function to help make it easier to code later and make sure everything works well. I made the routes of the website (url) proper to what they should go to. Just cleaned up the frontend. Backend I add a way to check the login and password with an sqlite database. THIS IS WITHOUT TEST CASES SO THEY NEED TO BE MADE AND CHECK IF THE CODE ACTUALLY WORKS OR NOT.